Rama bomba

+++ TPINT_GRUPO_2_LAB4/src/negocioImpl/CuentaNegocioImpl.java
@@ -1,8 +1,7 @@
 package negocioImpl;
 
 import java.sql.SQLException;
-import java.math.BigDecimal;
-import java.sql.SQLException;
+
 import java.util.ArrayList;
 
 import dao.CuentaDao;
@@ -162,7 +161,8 @@ public class CuentaNegocioImpl implements CuentaNegocio {
 	}
 	
 	@Override
-	public boolean ingresos(int idCuenta, BigDecimal montoACargar) {
+	public boolean ingresos(int idCuenta, Float montoACargar) {
+
 		return cuentaDao.ingresos(idCuenta, montoACargar);
 	}
 }
